+++ Toppick_App/lib/Products/Models/producto.dart
@@ -30,7 +30,7 @@ class Producto {
       json['descripcion'],
       json['precio'],
       json['tiempoPreparacion'],
-      (json['calificacion']==null)?0:json['calificacion'],
+      (json['calificacion']==null)?0:json['calificacion'].toDouble(),
       (json['urlImagen']==null)?"assets/img/pescadito.jpg":json['urlImagen'],
       json['categoria'],
       json['tipo']
